fix(ci): keep the test jobs inside the runner's memory
Both test jobs died with SIGTERM part-way through compiling the crate — no
rustc error, only "the runner has received a shutdown signal". The library
target and three integration binaries took `cargo test` from two heavy link
targets to six, and four concurrent codegen units over the polars/arrow
monomorphisations exhaust a 16 GB runner. Clippy passing over the same
--all-targets set is the tell: it emits metadata and never links. So CI drops
DWARF nothing reads, caps codegen on the two jobs that link, and stops the unit
job rebuilding the conformance binaries the conformance job already owns.

The harness had copied the status report, the log-page budget and the overload
retry hint out of the controller rather than calling them. The budget copy was
already 100 KiB looser than production, so the RP-22 page bound was checked
against a number the worker never uses. A harness that asserts against a
reimplementation of its subject can pass while the subject is broken, so those
three are now reachable from p2p and called.

test_fs asserted a raw directory listing, leaving it red on any macOS checkout
Finder had touched. Every production caller of `ls` already filters by name
shape, so the filtering belongs in the test and not in `ls`.

`conformance_metrics` read as "metrics about conformance" rather than
"conformance tests that touch process-global metrics", and `conformance.rs` was
a catch-all that would have silted up as the CT-2/4/9 classes land. Each file
is now named for what it tests, so the next class has an obvious home instead
of a default one.

The split of query_concurrency stays, but for the reason that actually forces
it: one test binary is one process, and the OB signals are process-global, so a
gauge assertion cannot sit beside tests that run queries for other reasons.
That is now stated where someone deciding where to put a new test will read it.

The harness self-test moves into the harness module — it guards the harness's
own honesty, so it should hold wherever the harness is used rather than in
whichever binary happened to be first.

Also adds the .gitattributes the repo never had. LFS was configured and the
index held correct pointers, but nothing routed *.parquet through the filter,
so git compared 356 MB of smudged content against 131-byte pointers and called
tests/data permanently modified — one `git add -A` away from inlining all of it
into the repo. The stored oids match the working tree exactly, so declaring the
filter reconciles the tree with no change to the data.
