+++ .gitattributes
@@ -1,0 +1,1 @@
+*.parquet filter=lfs diff=lfs merge=lfs -text
