+++ .github/workflows/tests.yaml
@@ -12,6 +12,9 @@ env:
   CARGO_TERM_COLOR: always
   PROTOC_VERSION: 33.3
   ARCH: linux-x86_64
+  # Nothing in CI reads a backtrace, and DWARF for the polars/arrow monomorphisations is the
+  # bulk of every binary here. The `test` profile inherits this.
+  CARGO_PROFILE_DEV_DEBUG: "0"
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.event_name == 'pull_request' && github.head_ref || github.ref }}
@@ -22,6 +25,11 @@ jobs:
     name: Test
     runs-on: ubuntu-latest
     if: github.event_name == 'push' || !github.event.pull_request.draft
+    # Each target here links ~300 MB of polars/arrow monomorphisations. Four codegen units
+    # at once exhausts a 16 GB runner: the agent dies with SIGTERM and no rustc error, only
+    # "the runner has received a shutdown signal".
+    env:
+      CARGO_BUILD_JOBS: "2"
     steps:
       - name: Cache protoc
         id: cache-protoc
@@ -47,8 +55,10 @@ jobs:
           lfs: true
       - uses: dsherret/rust-toolchain-file@v1
       - uses: Swatinem/rust-cache@v2
+      # Unit tier only; the conformance binaries are the `conformance` job's, and building
+      # them in both jobs doubles the heaviest link work for no extra coverage.
       - name: Run tests
-        run: cargo test --all-features
+        run: cargo test --all-features --lib --bins
 
   # MG-4: the PR conformance subset, kept as its own job so its wall clock is visible
   # against P-GATE-PR-TIME instead of hiding inside the unit-test run.
@@ -57,6 +67,9 @@ jobs:
     runs-on: ubuntu-latest
     timeout-minutes: 10 # P-GATE-PR-TIME
     if: github.event_name == 'push' || !github.event.pull_request.draft
+    # Three test binaries over the same heavy dependency graph — see the `test` job.
+    env:
+      CARGO_BUILD_JOBS: "2"
     steps:
       - name: Cache protoc
         id: cache-protoc
@@ -83,18 +96,22 @@ jobs:
           lfs: false
       - uses: dsherret/rust-toolchain-file@v1
       - uses: Swatinem/rust-cache@v2
-      - name: Run conformance suite
-        run: cargo test --test conformance --all-features
-      # Separate binary: the OB signals are process-global, so gauge assertions cannot
-      # share a process with other query-running tests.
-      - name: Run metrics conformance suite
-        run: cargo test --test conformance_metrics --all-features
+      # One binary per subject. query_concurrency is separate by necessity, not topic:
+      # the OB signals are process-global, so gauge assertions cannot share a process
+      # with tests that run queries for other reasons.
+      - name: Run conformance tier
+        run: |
+          cargo test --all-features \
+            --test e2e \
+            --test query_surface \
+            --test query_concurrency
       # A gap test stays ignored until its fix lands (MG-8: failing-test-first, never a
       # silent skip). Listing them keeps any quarantine visible in the job log.
       - name: List quarantined gap tests
         run: |
-          cargo test --test conformance --all-features -- --ignored --list
-          cargo test --test conformance_metrics --all-features -- --ignored --list
+          for t in e2e query_surface query_concurrency; do
+            cargo test --all-features --test "$t" -- --ignored --list
+          done
 
   fmt:
     name: Rustfmt
